# docs/openapi.json
{
  "components": {
    "schemas": {
      "Ability": {
        "description": "One ability slot on a Pokemon.",
        "properties": {
          "is_hidden": {
            "title": "Is Hidden",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slot": {
            "title": "Slot",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "slot",
          "is_hidden"
        ],
        "title": "Ability",
        "type": "object"
      },
      "Encounter": {
        "description": "A place a Pokemon appears, with the versions it appears in.",
        "properties": {
          "location_area": {
            "title": "Location Area",
            "type": "string"
          },
          "versions": {
            "items": {
              "type": "string"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "location_area",
          "versions"
        ],
        "title": "Encounter",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Health": {
        "description": "Liveness of the service.",
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "Health",
        "type": "object"
      },
      "Pokemon": {
        "description": "A Pokemon read from PokeAPI.",
        "properties": {
          "abilities": {
            "items": {
              "$ref": "#/components/schemas/Ability"
            },
            "title": "Abilities",
            "type": "array"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "types": {
            "items": {
              "$ref": "#/components/schemas/PokemonType"
            },
            "title": "Types",
            "type": "array"
          },
          "weight": {
            "title": "Weight",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "id",
          "types",
          "abilities",
          "height",
          "weight"
        ],
        "title": "Pokemon",
        "type": "object"
      },
      "PokemonType": {
        "description": "One type slot on a Pokemon.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slot": {
            "title": "Slot",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "slot"
        ],
        "title": "PokemonType",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PokeAPI Wrapper",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/pokemon/{name}": {
      "get": {
        "operationId": "read_pokemon_pokemon__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9-]+$",
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pokemon"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Pokemon"
      }
    },
    "/pokemon/{name}/encounters": {
      "get": {
        "operationId": "read_encounters_pokemon__name__encounters_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9-]+$",
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Encounter"
                  },
                  "title": "Response Read Encounters Pokemon  Name  Encounters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Encounters"
      }
    }
  }
}